Spilo: Add Load Balancer and nice DNS name for the PostgreSQL appliance.
The Load Balancer is created to route traffic to the current master. It is using a health check provided by Governor to determine the status of the master.
A CNAME DNS record is created for the load balancer.

To determine the Hosted Zone we currently ask the user to provide it during senza init.