Allow spilo  docker image to be specified at init.
As spilo resides in the open-source registry, we can
autodetect the latest image and put it into the generated
template. Alternatively, we allow a user to stick with the
previous behavior of specifying an image during create.

When getting the latest image, the logic is to fetch the
latest non-snapshot one before trying with snapshots.