Handle keyword parameters in create and print cmds
Positional and keyword parameters can be used together, much like
in Python.  A positional parameter must not follow one specified
with a keyword.

This also checks that no parameters is specified more than once
(either as positional or keyword) and that every parameter is
assigned a value.  Adding default parameter values should be easy
with this in place.