Add the ability to provide custom tags to all resources of the Cloud Formation stack.
To enable discovery of resources in aws we want to use tags to aid in finding our components.
The StackName will differ quite a lot between teams, as it is probably unique within 1 organization.

This patch adds the ability to provide a list of additional tags to be applied to all the resources.