Changing flow for setting traffic (#466)
* Adding the 'TargetGroup' for ALB inside the loop

When using an `ElasticLoadBalancerV2` the `TargetGroup` name was not appended on the actual ref but rather tried to be added to the list itself which causes a Python error to be thrown.

Looked like a copy and past error from the `ElasticLoadBalancer` version.

* Adding tests for ELB/ALB reference from ASG

These tests should catch the faulty behaviour which existed prior to the bugfix patch

* Stack updates are done on block in the end

When a stack has two ELBs/ALBs assigned setting traffic does not work as the second traffic update failes due to the stack already being updated for the first ELB domain. Now stack updates are done on block after all the changes have been processed.

* Treat successful stack update as change

* Fix the failing test_traffic test

It is just a quick fix and there should be some more elaborate test which really checks on the desired behaviour being executed not just the correct end result.