Fix issue with changed set traffic handling
The PR which introduced the changed handling of set traffic solving the UPDATE_IN_PROGRESS issue with stacks having multiple ELBs (and therefore multiple DNS names) did treat a stack object returned by CloudFormationStack.get_by_stack_name as if it would be the same object being returned by subsequent calls with the same stack_name. As this is not the case this leads to only the first update being actually executed.