Put nat_gateways into UserData if instance runs without public ip in DMZ (#477)
We have a use-case when we want to run instances in DMZ or public subnet
but without public ips. Elastic ip would be assigned later to one of the
instances.
Without public ip instance will not be able to initialize (download
docker, push logs, use AWS api, etc...). To solve this problem taupage
will create separate routing table for outgoing https requests and use a
custom nat gateway for that.