Extend spotinst elastigroup (#514)
This change tries to simplify the definition of Elastgroups by assuming some defaults like the native Auto Scaling Group counterpart.

It also adds support for the automatic subnet mapping from `Senza::StupsAutoConfiguration`

It assumes, for now, that the cloud accounts provisioned in Spotinst will be named "aws:<aws-account-id>". This is used for Spotinst cloud account resolution and it's needed to use a specific account instead of the default one.